Implement player shooting, Improved joystick
Player is able to shoot with two fire types single and auto. The aiming angle of the player has interpolation and will take some time to turn to the desired aiming direction.

Joystick now has many booleans, you can toggle when the joystick fails to get a direction vector whether it will reset the direction vector or keep it until it gets a valid direction. It also has booleans for being touched down, up, held and even if it is in the minimum deadzone.